+++ tests/integration/cqlengine/operators/test_where_operators.py
@@ -35,6 +35,7 @@ class TestWhereOperators(TestCase):
         check_lookup('LT', LessThanOperator)
         check_lookup('LTE', LessThanOrEqualOperator)
         check_lookup('CONTAINS', ContainsOperator)
+        check_lookup('LIKE', LikeOperator)
 
     def test_operator_rendering(self):
         """ tests symbols are rendered properly """
@@ -46,5 +47,6 @@ class TestWhereOperators(TestCase):
         self.assertEqual("<", six.text_type(LessThanOperator()))
         self.assertEqual("<=", six.text_type(LessThanOrEqualOperator()))
         self.assertEqual("CONTAINS", six.text_type(ContainsOperator()))
+        self.assertEqual("LIKE", six.text_type(LikeOperator()))