+++ CHANGELOG.rst
@@ -3,8 +3,12 @@
 
 Features
 --------
-
 * cqlengine: LIKE filter operator (PYTHON-512)
+
+Bug Fixes
+---------
+* Support retry_policy in PreparedStatement (PYTHON-861)
+
 
 3.12.0
 ======